pyproject.toml
[project]
name = "ecess-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.123.9",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
